hide admin stuff from non-admins
move some pages into pages dir
rearrange sidebar
