+++ webui/src/App.tsx
@@ -5,8 +5,8 @@ import ThemeProvider from "@mui/system/ThemeProvider";
 import { useEffect, useState } from "react";
 import { I18nextProvider } from "react-i18next";
 import i18n from "./i18n/i18n";
-import Login from "./Login";
-import { MainPageWithContext } from "./MainPage";
+import Login from "./pages/Login";
+import { MainPageWithContext } from "./pages/MainPage";
 import { checkIsLoggedIn } from "./rpc/backend";
 import { LoggedInResponse } from "./rpc/models";
 import { fold, isUndefined } from "./util/undefOr";
