+++ webui/src/pages/MainPage.tsx
@@ -1,3 +1,4 @@
+import AdminPanelSettingsIcon from "@mui/icons-material/AdminPanelSettings";
 import ApiIcon from "@mui/icons-material/Api";
 import GitHubIcon from "@mui/icons-material/GitHub";
 import KeyIcon from "@mui/icons-material/Key";
@@ -18,25 +19,56 @@ import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
-import { ReactNode, useCallback, useState } from "react";
+import { ReactNode, useCallback, useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { To, NavigateOptions, Outlet, useNavigate } from "react-router-dom";
-import { AppContext } from "./context";
-import { navigateToLogout } from "./rpc/backend";
-import { User } from "./rpc/models";
-import { useIsDesktop } from "./util/hooks";
+import { To, NavigateOptions, Outlet, useNavigate, createBrowserRouter, RouterProvider } from "react-router-dom";
+import { AppContext, useIsAdminUser } from "../context";
+import { navigateToLogout } from "../rpc/backend";
+import { User } from "../rpc/models";
+import { useIsDesktop } from "../util/hooks";
+import Administration from "./Administration";
+import { Api } from "./Api";
+import ChangePassword from "./ChangePassword";
+import { Search } from "./Search";
 
 const drawerWidth = 240;
 
+const Redirect = () => {
+  const nav = useNavigate();
+  useEffect(() => nav("/sd/search", { replace: true }), [nav]);
+  return <></>;
+};
+const createRouter = (admin: boolean) =>
+  createBrowserRouter([
+    { path: "/", element: <MainPage />, children: [{ index: true, element: <Redirect /> }] },
+    {
+      path: "/sd",
+      element: <MainPage />,
+      children: [
+        { index: true, element: <Redirect /> },
+        { path: "search", element: <Search /> },
+        { path: "search/:imageId", element: <Search /> },
+        { path: "api", element: <Api /> },
+        { path: "password", element: <ChangePassword /> },
+      ].concat(admin ? [{ path: "administration/*", element: <Administration /> }] : []),
+    },
+  ]);
+
 type Props = Readonly<{ user: User }>;
 
-export const MainPageWithContext = ({ user }: Props) => (
-  <AppContext.Provider value={{ currentUser: user }}>
-    <MainPage />
-  </AppContext.Provider>
-);
+export const MainPageWithContext = ({ user }: Props) => {
+  const isAdmin = user.admin;
+  const router = useMemo(() => createRouter(isAdmin), [isAdmin]);
+
+  return (
+    <AppContext.Provider value={{ currentUser: user }}>
+      <RouterProvider router={router} />
+    </AppContext.Provider>
+  );
+};
 
 const MainPage = () => {
+  const isAdminUser = useIsAdminUser();
   const isDesktop = useIsDesktop();
   const [drawerOpen, setDrawerOpen] = useState(false);
   const [t] = useTranslation();
@@ -101,9 +133,10 @@ const MainPage = () => {
           <Divider />
           <List>
             {navItem(t("common.changepw"), "password", <KeyIcon />)}
-            {menuItem(t("common.logout"), navigateToLogout, <LogoutIcon />)}
-            {navItem(t("admin.administration"), "administration", <KeyIcon />)}
+            {isAdminUser ? navItem(t("admin.administration"), "administration", <AdminPanelSettingsIcon />) : <></>}
           </List>
+          <Divider />
+          <List>{menuItem(t("common.logout"), navigateToLogout, <LogoutIcon />)}</List>
         </Box>
       </Drawer>
       <Box component="main" sx={{ flexGrow: 1, p: isDesktop ? 3 : 0, width: "100%" }}>
