+++ webui/src/util/undefOr.ts
@@ -2,5 +2,6 @@ export type UndefOr<T> = T | undefined;
 
 export const isDefined = <T>(t: UndefOr<T>): t is T => t !== undefined;
 export const isUndefined = <T>(t: UndefOr<T>): t is undefined => t === undefined;
+export const map = <T, R>(t: UndefOr<T>, f: (t: T) => R): UndefOr<R> => (isDefined(t) ? f(t) : undefined);
 export const fold = <T, R>(t: UndefOr<T>, ifUndef: () => R, ifDefined: (t: T) => R): R =>
   isDefined(t) ? ifDefined(t) : ifUndef();
